refactor: déplacer players vers features/players/
