+++ src/components/dialogs/index.ts
@@ -1,4 +1,6 @@
-// Player dialogs
-export { AddPlayerDialog } from './AddPlayerDialog';
-export { EditPlayerDialog } from './EditPlayerDialog';
-export { DeletePlayerDialog } from './DeletePlayerDialog';
+// All dialogs have been moved to feature directories.
+// This barrel is kept temporarily for compatibility.
+// Player dialogs → src/features/players/dialogs/
+// Game dialogs → src/features/games/dialogs/
+// Expansion dialogs → src/features/games/expansions/dialogs/
+// Character dialogs → src/features/games/characters/dialogs/
