+++ src/features/players/dialogs/DeletePlayerDialog.tsx
