+++ src/features/players/dialogs/index.ts
@@ -1,0 +1,3 @@
+export { AddPlayerDialog } from './AddPlayerDialog';
+export { EditPlayerDialog } from './EditPlayerDialog';
+export { DeletePlayerDialog } from './DeletePlayerDialog';
