+++ src/features/players/usePlayersPage.ts
@@ -1,6 +1,6 @@
 import { useState, useCallback } from 'react';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { playerApi } from '../services/api/playerApi';
+import { playerApi } from './playerApi';
 import { queryKeys } from '@/shared/services/api/queryKeys';
 import { useNavigationAdapter } from '@/shared/hooks/useNavigationAdapter';
 import type { Player, PlayerStatistics } from '@/types';
