+++ src/hooks/useDashboard.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { useQuery } from '@tanstack/react-query';
-import { playerApi } from '../services/api/playerApi';
+import { playerApi } from '@/features/players/playerApi';
 import { gameApi } from '@/features/games/gameApi';
 import { statsApi } from '@/shared/services/api/statsApi';
 import { queryKeys } from '@/shared/services/api/queryKeys';
